将Agent类重构并抽象化，并增加相应工厂类; 增加了两类Agent的对应实现; 实现Coordinator的两参数计算，并增加部分内部计算方法;

+++ BaseBuffer.cpp
@@ -9,5 +9,5 @@ BaseBuffer::BaseBuffer() {}
 BaseBuffer::BaseBuffer(String id, String bdType, long rq) :boardId(id), boardType(bdType), rqId(rq) {}
 
 String BaseBuffer::getBoardId() { return this->boardId; }
-int BaseBuffer::getReqId() { return this->rqId; }
+long BaseBuffer::getReqId() { return this->rqId; }
 String BaseBuffer::getBoardType() { return this->boardType; }
